cpuidle: Make pm_qos notifier callback more efficient
Currently, the cpumask API is used to determine which CPUs need to have
an IPI sent, but the cpumask API uses the bitops API under the hood,
which is quite inefficient. In kernels older than 4.19, the bitops API
uses spin locks to guarantee atomicity, which is a major blow to
efficiency in a path as hot as cpuidle_latency_notify().

The cpumask API doesn't need to be used at all in fact, since we're
guaranteed to have a total number of CPUs that can fit within a word.

Replace the cpumask usage with raw bitwise operations to make the pm_qos
notifier callback more efficient. While we're at it, check to make sure
the mask isn't empty when arch_send_call_function_ipi_mask() receives
it, since arch_send_call_function_ipi_mask() doesn't check for empty
CPU masks and will do some work when given one.

* 3.18 without cpu_isolated_mask

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: zeta96